requirements.txt
Flask==2.3.2
Pillow==9.5.0
gunicorn==20.1.0
